refactor(emitter): split commonjs transform tests
